requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
beautifulsoup4==4.6.0
certifi==2017.11.5
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.2.1
cssselect==1.0.3
defusedxml==0.5.0
docopt==0.6.2
fake-useragent==0.1.10
Flask==0.12.2
Flask-Login==0.4.1
geoip2==2.8.0
gitdb2==2.0.3
GitPython==2.1.9
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.10
lxml==4.1.1
MarkupSafe==1.0
maxminddb==1.3.0
parsel==1.4.0
pipreqs==0.4.9
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pycurl==7.43.0.1
PyDispatcher==2.0.5
pymongo==3.6.1
pyOpenSSL==17.5.0
pyquery==1.4.0
PySocks==1.6.8
pyspider==0.3.9
pywin32==221
queuelib==1.5.0
redis==2.10.6
requests==2.18.4
Scrapy==1.5.0
selenium==3.11.0
service-identity==17.0.0
six==1.11.0
smmap2==2.0.3
tblib==1.3.2
tornado==5.0.1
Twisted==17.9.0
u-msgpack-python==2.4.1
urllib3==1.22
w3lib==1.19.0
Werkzeug==0.14.1
win-inet-pton==1.0.1
WsgiDAV==2.2.4
yarg==0.1.9
zope.interface==4.4.3
